backend/requirements.txt
Django==4.2.4
django-cors-headers == 3.13.0
djangorestframework == 3.14.0
gunicorn == 20.1.0